requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
boto3==1.34.96
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic[dotenv]==2.7.0
pytz==2024.1

